Fix: useDetectWebView hook client환경에서만 실행되도록 수정
